Feature/NTA_Restapi
Moved Get and GetById methods to a generic Common Service interface.
Added two specific interfaces for Transaction and ResourceManager which extends Common Service interface.
Updated the implementation of the service classes to implement Common Service interface.
Updated the REST API to return bad request for a entity that does not exists in the system.